+++ client/dom/table.js
@@ -1,4 +1,4 @@
-import { select } from 'd3'
+import { select, style } from 'd3'
 import { defaultcolor } from '../shared/common'
 
 /*
@@ -11,6 +11,7 @@ div = d3-wrapped holder
 columns = [ {label} ]
 	each element is an object describing a column
 	label: str, the text to show as header of a column
+	width: str, column width
 
 rows = [ [] ]
 	each element is an array of cells for a row, with array length must matching columns.length
@@ -30,87 +31,132 @@ rows = [ [] ]
 			{url/html/value}, {}, ...
 		]
 	}
+style = {}
+
+	max_width: str, the max width of the table, if not provided is set to 90vw
+	max_height: str, the max height of the table, if not provided is set to 50vh
+	row_height: str, the height of the row
+
+buttons = [ {button} ]
+	Each element is an object describing a button:
+	text: str, the text to show in the button
+	callback: function, the function to be called when the button is clicked
+
+noButtonCallback = (index, node) => {}
+	Function that will be called when a row is selected	
+
+singleMode = false, boolean
+	Specifies if a radio button should be rendered instead
+	
 */
-export async function renderTable({ columns, rows, div, max_width = '90vw', max_height = '50vh', buttons }) {
+export async function renderTable({ columns, rows, div, style = {}, buttons, noButtonCallback, singleMode = false }) {
 	const numColumns = columns.length
 
-	// create a Parent Div element to which the header and sample grid will be appended as divH and divS.
+	// create a Parent Div element to which the header and sample table will be appended as divH and divS.
 	const parentDiv = div
-		.append('div')
-		.style('overflow', 'scroll')
-		.style('scrollbar-width', 'none')
-		.style('width', '98%')
+		.style('padding', '5px')
 		.style('background-color', 'white')
-		.attr('class', 'sjpp_grid_container')
-		.style('grid-template-columns', `1.5vw ${buttons ? '2vw' : ''} repeat(${numColumns}, auto)`)
-		.style('max-width', max_width)
-		.style('max-height', max_height)
+		.append('table')
+		.style('display', 'block')
+		.style('background-color', 'white')
+		.style('max-width', style.max_width ? style.max_width : '90vw')
 
 	// header div
-	const divH = parentDiv.append('div').style('display', 'contents')
-	divH
-		.append('div')
-		.attr('class', 'sjpp_grid_item sjpp_grid_header')
-		.text('#')
-	if (buttons) {
-		const cell = divH.append('div').attr('class', 'sjpp_grid_item')
-		const checkboxH = cell
-			.append('input')
-			.attr('type', 'checkbox')
-			.on('change', () => {
-				table.selectAll('input').property('checked', checkboxH.node().checked)
-				enableButtons()
-			})
+	const divH = parentDiv
+		.append('thead')
+		.style('display', 'table')
+		.style('table-layout', 'fixed')
+		.style('width', '100%')
+		.append('tr')
+	if (style.show_lines) {
+		divH
+			.append('th')
+			.attr('class', 'sjpp_table_header')
+			.text('#')
+			.style('width', '15px')
+	}
+
+	if (buttons || noButtonCallback) {
+		const cell = divH
+			.append('th')
+			.attr('class', 'sjpp_table_header')
+			.style('width', '20px')
+		if (!singleMode) {
+			const checkboxH = cell
+				.append('input')
+				.attr('type', 'checkbox')
+				.on('change', () => {
+					table.selectAll('input').property('checked', checkboxH.node().checked)
+					enableButtons()
+				})
+		}
 	}
 
 	// header values
 	for (const c of columns) {
-		divH
-			.append('div')
+		const th = divH
+			.append('th')
 			.text(c.label)
-			.attr('class', 'sjpp_grid_item sjpp_grid_header')
+			.attr('class', 'sjpp_table_item sjpp_table_header')
+		if (c.width) th.style('width', c.width)
 	}
 
-	// sample values
-	// iterate over each row in rows and create a div for each row that has a grid layout similar to the header grid.
-	const table = parentDiv.append('div').style('display', 'contents')
+	const table = parentDiv
+		.append('tbody')
+		.style('display', 'block')
+		.style('max-height', style.max_height ? style.max_height : '50vh')
+		.style('overflow', 'scroll')
 
 	for (const [i, row] of rows.entries()) {
-		const rowGrid = table.append('div')
-		rowGrid.attr('class', 'sjpp_grid_row_wrapper')
-		const lineDiv = rowGrid
-			.append('div')
-			.text(i + 1)
-			.style('font-size', '0.8rem')
-			.style('color', defaultcolor)
-			.attr('class', 'sjpp_grid_item')
-			.style('background-color', i % 2 == 0 ? 'rgb(237, 237, 237)' : 'white')
-
-		if (buttons) {
-			const checkbox = rowGrid
-				.append('div')
-				.attr('class', 'sjpp_grid_item')
-				.style('background-color', i % 2 == 0 ? 'rgb(237, 237, 237)' : 'white')
-				.style('float', 'center')
-				.append('input')
-				.attr('type', 'checkbox')
-				.attr('value', i)
-				.on('change', () => enableButtons())
+		let checkbox
+		const rowtable = table
+			.append('tr')
+			.attr('class', 'sjpp_row_wrapper')
+			.style('display', 'table')
+			.style('table-layout', 'fixed')
+			.style('width', '100%')
+		if (buttons || noButtonCallback)
+			rowtable.on('click', e => {
+				if (e.target !== checkbox.node()) {
+					if (singleMode)
+						//not a checkbox
+						checkbox.node().checked = true
+					else checkbox.node().checked = !checkbox.node().checked
+					checkbox.dispatch('change')
+				}
+			})
+		if (style.show_lines) {
+			const lineDiv = rowtable
+				.append('td')
+				.text(i + 1)
+				.style('width', '15px')
+				.style('font-size', '0.8rem')
+				.style('color', defaultcolor)
 		}
 
-		// each row comprises of cell and each cell has values that will get appended to div elements of the rowGrid stored in td.
+		if (buttons || noButtonCallback) {
+			checkbox = rowtable
+				.append('td')
+				.style('width', '20px')
+				.style('float', 'center')
+				.append('input')
+				.attr('type', singleMode ? 'radio' : 'checkbox')
+				.attr('name', 'select')
+				.attr('value', i)
+				.on('change', () => {
+					if (buttons) enableButtons()
+					else noButtonCallback(i, checkbox.node())
+				})
+		}
+
 		for (const [colIdx, cell] of row.entries()) {
-			const td = rowGrid
-				.append('div')
-				.attr('class', 'sjpp_grid_item')
-				.style('background-color', i % 2 == 0 ? 'rgb(237, 237, 237)' : 'white')
+			const td = rowtable.append('td').attr('class', 'sjpp_table_item')
+			const column = columns[colIdx]
+			if (column.width) td.style('width', column.width)
 
-			// if index of each row is even then the background of that row should be grey and also add hovering in yellow.
-
-			// if cell has values then append those values in new divs on td which is stored in d.
 			if (cell.values) {
 				for (const v of cell.values) {
-					// if those values have url in them then tag it to the sample name/id otherwise just append the value of that cell onto the div
+					// if those values have url in them then tag it to the sample name/id otherwise just append the value of that cell onto the td
 					if (v.url) {
 						td.append('a')
 							.text(v.value)
@@ -137,9 +183,9 @@ export async function renderTable({ columns, rows, div, max_width = '90vw', max_
 	if (buttons) {
 		const footerDiv = div
 			.append('div')
+			.insert('div')
 			.style('display', 'inline-block')
 			.style('float', 'right')
-			.style('margin', '5px 5px')
 
 		for (const button of buttons) {
 			const values = []
@@ -148,7 +194,7 @@ export async function renderTable({ columns, rows, div, max_width = '90vw', max_
 				.append('button')
 				.attr('disabled', true)
 				.text(button.text)
-				.style('margin-right', '10px')
+				.style('margin', '10px 10px 0 0')
 				.on('click', e => {
 					const checkboxs = table.selectAll('input:checked')
 					if (!checkboxs.empty()) {