+++ client/plots/sampleScatter.js
@@ -119,6 +119,7 @@ class Scatter {
 		const data = await this.app.vocabApi.getScatterData(reqOpts)
 		if (data.error) throw data.error
 		if (!Array.isArray(data.samples)) throw 'data.samples[] not array'
+
 		this.shapeLegend = new Map(Object.entries(data.shapeLegend))
 		this.colorLegend = new Map(Object.entries(data.colorLegend))
 
@@ -526,7 +527,12 @@ function setRenderers(self) {
 				.text(`List ${self.selectedItems.length} samples`)
 				.on('click', event => {
 					self.dom.tip.hide()
-					showTable({ name: 'Selected samples', items: self.selectedItems }, event.clientX, event.clientY, true)
+					showTable(
+						{ name: 'Group ' + (self.config.groups.length + 1), items: self.selectedItems },
+						event.clientX,
+						event.clientY,
+						true
+					)
 				})
 
 			menuDiv
@@ -810,13 +816,17 @@ function setRenderers(self) {
 			rows.push(row)
 		}
 		self.dom.tip.clear()
-		const headerDiv = self.dom.tip.d.append('div')
-		headerDiv
-			.insert('div')
-			.html('&nbsp;' + group.name)
-			.style('display', 'inline-block')
-			.style('font-weight', 'bold')
-			.style('margin-top', '5px')
+		const headerDiv = self.dom.tip.d.append('div').style('margin-top', '5px')
+
+		headerDiv.insert('label').html('&nbsp;Group: ')
+		const groupInput = headerDiv
+			.insert('input')
+			.attr('value', group.name)
+			.on('change', e => {
+				const value = groupInput.node().value
+				if (value) group.name = value
+				else groupInput.node().value = group.name
+			})
 		const tableDiv = self.dom.tip.d.append('div')
 		const buttons = []
 		if (addGroup) {
@@ -826,34 +836,37 @@ function setRenderers(self) {
 					const items = []
 					for (const i of indexes) items.push(self.selectedItems[i])
 					self.config.groups.push({
-						name: `Group ${self.config.groups.length + 1}`,
+						name: group.name,
 						items,
 						index: groups.length
 					})
+					console.log(self.config.groups)
 					self.app.dispatch({ type: 'plot_edit', id: self.id, config: { groups: self.config.groups } })
 				}
 			}
 			buttons.push(addGroup)
-		}
-		const deleteSamples = {
-			text: 'Delete samples',
-			callback: indexes => {
-				group.items = group.items.filter((elem, index, array) => !(index in indexes))
-				showTable(group, x, y, addGroup)
+		} else {
+			const deleteSamples = {
+				text: 'Delete samples',
+				callback: indexes => {
+					group.items = group.items.filter((elem, index, array) => !(index in indexes))
+					showTable(group, x, y, addGroup)
+				}
 			}
+			buttons.push(deleteSamples)
 		}
-		buttons.push(deleteSamples)
 		renderTable({
 			rows,
 			columns,
 			div: tableDiv,
+			style: { show_lines: true, max_width: '35vw', max_height: '35vh' },
 			buttons
 		})
 
 		self.dom.tip.show(x, y)
-
+		//scroll(x, y)
 		function formatCell(column, name = 'value') {
-			let dict = {}
+			let dict = { width: '10vw' }
 			dict[name] = column
 			return dict
 		}
@@ -862,7 +875,6 @@ function setRenderers(self) {
 
 function setInteractivity(self) {
 	self.mouseover = function(event) {
-		if (self.lassoOn) return
 		if (event.target.tagName == 'path') {
 			const d = event.target.__data__
 			if (!d) return